Reducing the amount of warnings in Debug/x86 build (/ZI+SEH incompatibility and signed/unsigned warning). (#127)

+++ libs/StateSync-pinvoke-cpp/src/Key-pinvoke.cpp
@@ -32,7 +32,7 @@ Microsoft_MixedReality_Sharing_StateSync_Key_view(intptr_t handle,
   if (auto* key = bit_cast<const Key*>(handle)) {
     auto view = key->view();
     // TODO: ensure that we never allow keys larger than INT_MAX
-    assert(view.size() < std::numeric_limits<int>::max());
+    assert(view.size() < static_cast<size_t>(std::numeric_limits<int>::max()));
     *out_size = static_cast<int>(view.size());
     return view.data();
   }
